Fix an issue with our docker build
The issue is our `.dockerignore` file blocks anyting from the `test/` directory from getting into the image. But we have now built a controller (`/admin/test/bill-runs/generate`) which depends on a number of things in the test folder.

Loading a newly built image and trying to run it exposed the issue.

A check of the logs shows we have yet to try and deploy an image built after [Add test only bill run generator](https://github.com/DEFRA/sroc-charging-module-api/pull/199) was merged in. But if we did we would expect it to error on startup with

```
uncaught exception Error: Cannot find module '../../../../test/support/generators'
Require stack:
- /home/node/app/app/controllers/admin/test/test_bill_run.controller.js
- /home/node/app/app/controllers/index.js
- /home/node/app/app/routes/airbrake.routes.js
- /home/node/app/app/routes/index.js
- /home/node/app/app/plugins/router.plugin.js
- /home/node/app/app/plugins/index.js
- /home/node/app/server.js
    at Function.Module._resolveFilename (internal/modules/cjs/loader.js:831:15)
    at Function.Module._load (internal/modules/cjs/loader.js:687:27)
    at Module.require (internal/modules/cjs/loader.js:903:19)
    at require (internal/modules/cjs/helpers.js:74:18)
    at Object.<anonymous> (/home/node/app/app/controllers/admin/test/test_bill_run.controller.js:4:30)
    at Module._compile (internal/modules/cjs/loader.js:1015:30)
    at Object.Module._extensions..js (internal/modules/cjs/loader.js:1035:10)
    at Module.load (internal/modules/cjs/loader.js:879:32)
    at Function.Module._load (internal/modules/cjs/loader.js:724:14)
    at Module.require (internal/modules/cjs/loader.js:903:19) {
  code: 'MODULE_NOT_FOUND',
  requireStack: [
    '/home/node/app/app/controllers/admin/test/test_bill_run.controller.js',
    '/home/node/app/app/controllers/index.js',
    '/home/node/app/app/routes/airbrake.routes.js',
    '/home/node/app/app/routes/index.js',
    '/home/node/app/app/plugins/router.plugin.js',
    '/home/node/app/app/plugins/index.js',
    '/home/node/app/server.js'
  ]
}
```