No long push git commit tag to docker hub
After a re-think, we are going to stick to our existing strategy with Docker hub. We will still automatically build `main` as `latest`. We are also going to hook up DockerHub to automatically build the tags we create.

But the incremental builds we are going to leave off and we will instead get into the practise of generating releases when new endpoints are added or major changes are made, which includes release notes, and informing the WRLS that they are available.