Fix docker build issues due to bill run generator
We recently added a [test only bill run generator](https://github.com/DEFRA/sroc-charging-module-api/pull/199) which can be used to generate bill runs made up of different types of invoices.

To make it as quick as possible, and to avoid spamming the rules service with lots of requests we made use of some of our `test/support` functionality. Mainly, we mock the rules service using [nock](https://github.com/nock/nock) and use our helpers and fixtures as part of doing that.

What we had overlooked, and had not spotted till now is that any attempt to start the app within latest images would error.

```
uncaught exception Error: Cannot find module '../../../../test/support/generators'
Require stack:
- /home/node/app/app/controllers/admin/test/test_bill_run.controller.js
- /home/node/app/app/controllers/index.js
- /home/node/app/app/routes/airbrake.routes.js
- /home/node/app/app/routes/index.js
- /home/node/app/app/plugins/router.plugin.js
- /home/node/app/app/plugins/index.js
- /home/node/app/server.js
    at Function.Module._resolveFilename (internal/modules/cjs/loader.js:831:15)
    at Function.Module._load (internal/modules/cjs/loader.js:687:27)
    at Module.require (internal/modules/cjs/loader.js:903:19)
    at require (internal/modules/cjs/helpers.js:74:18)
    at Object.<anonymous> (/home/node/app/app/controllers/admin/test/test_bill_run.controller.js:4:30)
    at Module._compile (internal/modules/cjs/loader.js:1015:30)
    at Object.Module._extensions..js (internal/modules/cjs/loader.js:1035:10)
    at Module.load (internal/modules/cjs/loader.js:879:32)
    at Function.Module._load (internal/modules/cjs/loader.js:724:14)
    at Module.require (internal/modules/cjs/loader.js:903:19) {
  code: 'MODULE_NOT_FOUND',
  requireStack: [
    '/home/node/app/app/controllers/admin/test/test_bill_run.controller.js',
    '/home/node/app/app/controllers/index.js',
    '/home/node/app/app/routes/airbrake.routes.js',
    '/home/node/app/app/routes/index.js',
    '/home/node/app/app/plugins/router.plugin.js',
    '/home/node/app/app/plugins/index.js',
    '/home/node/app/server.js'
  ]
}
```

The reason is our `.dockerignore` file blocks anything from the `test/` directory from getting into the image during the build. But we have added `TestBillRunController` which depends on `BillRunGenerator` located in `test/` directory. and `BillRunGenerator` in turn relies on fixtures and helpers in `test/`.

So, to get this building we need to update `.dockerignore` to allow the `test/support` and all its contents to be copied to the image during build.

We also need to move the `nock` dependency from `devDependencies`. Again, as a dev only dependency it's not available when we build our **production** level docker image. So, we need to move it to `dependencies`.