Add support for automated builds in Docker Hub
https://trello.com/c/jHORNdnQ

We have been asked to start making available this project's Docker images so the WRLS team have a means to start working with the new version and migrating their code to use it.

We have done this before with [version 1 of the API](https://github.com/defra/charging-module-api) by creating a [repository on Docker Hub](https://hub.docker.com/r/environmentagency/charging-module-api) and pushing each release image to it.

This was fine for v1 where there was months between releases. But both teams are actively working on switching away from V1, so releases should now come think and fast as new endpoints are added and fixes and improvements are made.

Docker Hub can be linked to your repo which then allows [Automated builds](https://docs.docker.com/docker-hub/builds/) to be enabled. This means every change to `main` will kick off an automated build on Docker Hub.

- we get feedback integrated with our project as to whether the image can still be built
- we can make available the very latest changes in a Docker image to WRLS

On top of that we can also trigger Docker Hub to build every time we push a new tag to the repo. This means we no longer have to manually pull, re-tag and push images from Artifactory.

So this changes makes all the changes necessary to support integration with Docker Hub.