Add Dockerhub build hook
Without repeating too much of what is in the new README, we have to override the default Docker hub build command to allow us to set the git commit and image name as build args.

To do this we have to add the file `hooks/build` containing the actual build command docker hub should use for automated builds.