Update existing auth systems seed
We'll now grab the `SYSTEM_CLIENT_ID` env var if its there in our config. We use this in the authorised_systems seed file to determine whether to seed the system user or not.

Note - Manually tested seeding with the env var present and not present to confirm it works as expected in both scenarios.