Make transaction tracking properties consistent
We spotted during a review that the properties which track the number of debit and credit transactions at bill run level don't share the same name as the ones at invoice, and licence level.

At the bill run level they are

- `creditLineCount`
- `creditLineValue`
- `debitLineCount`
- `debitLineValue`

At invoice and licence level they are

- `creditCount`
- `creditValue`
- `debitCount`
- `debitValue`

Behind the scenes they are all the same; the short name is used. But after discussion the dev team is agreed that the longer `creditLineCount` is clearer and should be used at both database and response level.

This change covers everything needed to get us to that position.