Start handling new bill run requests
We want to handle attempts to edit bill runs that cannot be edited in the service. As a pre-cursor to that though we need to handle `POST` requests to create new bill runs.

An initial change we can make is to only consider requests bill run related if the path includes `/bill-runs/`, the closing slash being critical. This is because create bill run requests will come to `/bill-runs` so can be safely ignored.