Revert previous change and update test route
On review it made better sense to change the test bill runs controller route than switch our simple regex for a muc more complex one.

Now the test bill run controller is more consistent with the real one. A `POST` request to `/bill-runs` is sufficient to tell us a new bill run needs creating. We don't need `/generate` in the path. And we can revert back to our simpler to understand regex in the new `RequestBillRunService`.