Add admin test view transaction endpoint
https://trello.com/c/0m5idNGk

This is actually a re-think of the [Add stand-in view bill run transaction endpoint](https://github.com/DEFRA/sroc-charging-module-api/pull/211). [Version 1](https://github.com/defra/charging-module.api) of the API has view bill run transaction endpoint though discussions with the WRLS team have confirmed it's not used by them.

The main driver for replicating it in Version 2 of the API has actually been to support testing; our lead QA @anwarisse1 needs a view of the transaction to confirm everything is populated as expected in a range of scenarios.

So, we started PR #211 with the intent of providing a client focused endpoint, even though WRLS re-confirmed they wouldn't use it and knowing it was just a priority for testing. They also confirmed they only want to see a very small subset of the fields Version 1 currently returns. Again, if we were truly building this endpoint for the client we'd focus on just the fields needed, which would hinder testing.

Having taken a step back, we've decided that at this time what is really needed is an endpoint where the focus is _testing_ and not a potential client system.

Because of this, we're doing this change which adds a view transaction endpoint to our `/admin/test` area of the API. Plus, like the regime and authorised system endpoints we'll just return everything the database returns by way of the `JsonPresenter`.

Should a 'true' client facing view transaction endpoint be needed, we can continue with the work in PR #211.