Add Bill Run Invoice Presenter
https://trello.com/c/r9UClZob

This is the second part of adding a view bill run invoice endpoint to the service; adding the presenter.

The presenter will return the invoice detail plus licence and transaction detail as WRLS have requested. We also have a hit list of fields WRLS would like to see for each transaction.