Amend remove invoice route
https://trello.com/c/VV4vIldJ

This is an amendment to the change [Initial Delete Invoice controller and endpoint](https://github.com/DEFRA/sroc-charging-module-api/pull/219).

We should have spotted at the time that the route is not aligned with current agreements with the WRLS team. The convention going forward is if something is related to a bill run, actions for it come under the `/bill-runs/{billRunId}` path.

This change just tweaks the standin endpoint to follow this convention.