Add get bill run plugin (#204)
There are a number of places where we need to query the DB for a matching bill run based on the bill run ID provided. In fact, the majority of our routes are on `/v2/{regime}/bill-runs`.

This has led to duplication in a number of services that need to grab the bill run as part of their work and in some cases more than once in the same request.

When we encounter something like this (`Regime` being the primary example) we move the functionality to a plugin. This will automatically be executed as part of all requests and means we can centralise the logic for getting and performing initial validation on bill runs.

This change adds the new plugin.

* Notes

Something our existing code didn't handle was when a request related to a bill run came in, but it wasn't linked to the requested regime.

This is important because our authorisation logic is based on the regime; can the client system perform said action for the requested regime.

If we didn't start looking at both regime and bill run client systems could circumvent this. For example, send a request to view a bill run using a regime they do have access to, but specify a bill run linked to a regime they don't. Currently, we would allow this to succeed when we shouldn't.

---

We also updated the existing test bill runs generator route to use a path that would

- mean it's not picked up by this plugin
- is consistent with the proper create a bill run endpoint

A `POST` request to `/bill-runs` is sufficient to tell us a new bill run needs creating. We don't need `/generate` in the path.