Remove presenter test
As per our current convention, if a presenter is only renaming fields we don't see the value in maintaining a unit test for it, especially as it often means coming up with a large dummy data set.

So, this commit removes the test we were using to confirm the presenter is working as expected.