Remove test
Continuing with our convention of not including tests for 'simple' presenters.

In this case there is some logic tucked away into the presenter. But it is our intention to retain tests for the ViewInvoicePresenter which will just duplicate what we are doing here. This is why we are again happy to delete these tests now.