Update implementation
We learnt how to create an array of 'presented' data in [Add view licence presenter](https://github.com/DEFRA/sroc-charging-module-api/pull/230) so we implement the exact same technique hear.

We also can test properly for 'presented' values so we are able to update our tests to match what will actually be returned.