Fix inconsistent test bill run controller name
The PR [Add test only bill run generator](https://github.com/DEFRA/sroc-charging-module-api/pull/199) added a new endpoint that allows us to generate bill runs for testing.

There is no issue with the logic, it's just an issue with the naming of the controller. Our convention is to make them plural but `TestBillRunController` is singular.

This is a quick fix to stop our eyes twitching when we see it!