Fix broken validation service test
The reason it was broken before is because the instance of `BillRunModel` we were passing into the service was set before the call to `CreateTransactionService`. We could have [re-fetched](https://vincit.github.io/objection.js/api/model/instance-methods.html#query) the instance and then passed it in.

But we know what the service (and the underlying `$empty()` method) are actually checking for us whether any of the line counts are greater than 0. So, instead we have vastly simplified the test by just doing that, and removing the need for any transactions to actually be added, and the stubbing of the rules service required to support that.