Fix generate bill run handler and add note
Spotted after making the previous changes that we had forgotten to update the call to `GenerateBillRunService` in the controller. So, it was still passing in an ID.

Initially this caused concern because we had controller tests still passing. We then realised this is expected behaviour. The bill run had passed validation so we immediately return a 204 response. However, we kick off the generate process asynchronously which means it can still fail after we have returned the response. In the hopes of not getting panicked again, we have added an extra note to the test title.