Hold off ViewBillRunService
The `ViewBillRunService` currently assumes it has to get the bill run. But, as it only needs to provide a subset of the data it amends the query it uses to find the bill run, to pull through the linked invoices and licences, but only key fields for them.

We feel that needs a bit more thought so will hold off doing it in this PR. Which means we're done here for now!