Add references to licences and transactions tables (#228)
We intend to use cascading deletes to make the Delete Invoice Service automatically remove licences and transactions when an invoice is deleted. To do this we need to add a reference from the transactions table to the licences table, and from the licences table to the invoices table. However, doing this causes issues with our tests, as we can no longer just insert a single licence to test; it would need a corresponding invoice as well. And likewise for transactions, which would need a licence (which in turn would need an invoice).

This changes adds the references to the licences and transactions tables, and makes changes to the test helpers as follows:

* LicenceHelper is introduced which can be used to create a licence. If an invoice id isn't specified when creating the licence, it will call InvoiceHelper to create one.
* TransactionHelper will now create a licence for a transaction if one isn't specified. It does this using LicenceHelper so an invoice will also be created if required.