Add view transaction presenter (#229)
https://trello.com/c/r9UClZob

As part of working towards having a view invoice endpoint, we need to add a `ViewTransactionPresenter`. Eventually, the view invoice endpoint will return a response that includes invoice, licence and transaction details. We'll need a presenter for each level to enable that though.

So, this change adds a `ViewTransactionPresenter` which is intended to be used by a `ViewLicencePresenter`, which in turn will be used by a `ViewInvoicePresenter`.