Remove redundant bill run 'gets' and 'checks' (#233)
PR [Add get bill run plugin](https://github.com/DEFRA/sroc-charging-module-api/pull/204) added a new plugin to handle getting and checking the bill run for all `/bill-runs/` related requests. Now we have one place that grabs the bill run, checks it's linked to the requested regime and if relevant, can be changed by the request.

It won't cover all the checks a service needs to make. But it does mean we can delete any calls to grab the bill run or checks that duplicate what the bill run does from our services.

This change covers hunting for those duplications and removing them.

** Notes

* Rename bill run validation service

We can drop a number of the checks in the service thanks to the plugin. This means those that remain are specific to generating a bill run so it makes sense to rename the service.

* Fix broken validation service test

The reason it was broken is that the instance of `BillRunModel` we were passing into the service was set before the call to `CreateTransactionService`. We could have [re-fetched](https://vincit.github.io/objection.js/api/model/instance-methods.html#query) the instance and then passed it in.

But we know what the service (and the underlying `$empty()` method) are actually checking for is whether any of the line counts are greater than 0. So, instead, we have vastly simplified the test by just setting one of those counts, and removing the need for any transactions to actually be added, or the stubbing of the rules service required to support it.

* Fix generate bill run handler and add note

Spotted after making the previous changes that we had forgotten to update the call to `GenerateBillRunService` in the controller. So, it was still passing in an ID.

Initially, this caused concern because we had controller tests still passing. We then realised this is expected behaviour. The bill run had passed validation so we immediately return a 204 response. However, we kick off the generate process asynchronously which means it can still fail after we have returned the response. In the hopes of not getting panicked again, we have added an extra note to the test title.