Fix bill run $editable() method (#234)
We have just spotted that the `BillRunModel.$editable()` is wrong. It is assessing a bill run as editable as long as its status is either `initialised` or `summarised`. However, when we finally added the [generate bill run](https://github.com/DEFRA/sroc-charging-module-api/pull/181) functionality we instead chose to use `generated` not `summarised`.

This change fixes the issue.