Add stand-in approve bill run endpoint
https://trello.com/c/SmiI8YaG

This is the first step in adding a `PATCH /v2/{regime}/bill-run/{id}/approve` endpoint to the API.

The roots of this project are based in the work done on the [sroc-tcm-admin](https://github.com/DEFRA/sroc-tcm-admin) service. In that service individual transactions would need to be approved first at which point they would be included in the next 'bill run' (though it doesn't use that term).

As this API has been developed to serve the WRLS service first, 'bill runs' and creating them first then adding transactions to them came to the fore. But the idea of needing to 'approve' transactions has remained.

The overall need is to add an endpoint that allows the transactions linked to a bill run to be mass approved. This change gets things started with a stand in endpoint ready to hook up to a service.