Fix issue with summarise invoice and credit issues
It's a tough one to spot, but the issue was that the queries being run to form the basis of the calculations for the invoices and credit notes was being done 'outside the transaction'.

Inserts and updates and still persisted within the 'context' of a transaction. It's DB magic that then persists everything 'for real' when the transaction is committed.

So, the following example (using pseudo code to keep things simple) would work

```
await myDatabase.transaction(async (trx) => {
  let person = { name: 'Alan', age: 32 } // shut your face I am!
  myDatabase.insert(trx, person)

  person = myDatabase.select(trx, 'people', "WHERE name = 'Alan'")
  console.log(person.age) // 32

  myDatabase.update(trx, 'people', "WHERE name = 'Alan'", 'SET age = (age + 10)')

  person = myDatabase.select(trx, 'people', "WHERE name = 'Alan'")
  console.log(person.age) // 42 - boo!
})
```

A `SELECT` based query which follows an `UPDATE` or `INSERT` _will_ see the new data but only if the query is done _within the transaction_.