Break up bill run and invoice tests
It is true that we get a little less code and a bit more speed if we combine the tests. But as tests can (and IMHO should) serve as documentation it is fine to sacrifice a little time for focus and clarity.

Hence, switched back to a format where the 2 are split out. I'm hoping (😰) they were grouped up just to help diagnose the original issue and not because of a competing belief system!