Add view bill run invoice service
https://trello.com/c/r9UClZob

As part of the new view bill run invoice endpoint, we need a service that manages the process which the handler can call on.

It needs to grab and validate the invoice (for which we'll use the new `FetchAndValidateBillRunInvoiceService`) then combine the result with the new `ViewInvoicePresenter`.

This change adds that service!