Create Delete Invoice Service (#221)
https://trello.com/c/VV4vIldJ

The next stage of setting up Delete Invoice Transactions is to create a Delete Invoice Service.

The service reads the stats of the invoice it's called on (debit/credit counts and values, etc.) and patches the db to subtract these values from the invoice's bill run.

This change also amends the licence and transaction tables so that deleting an invoice cascades down to deleting all its licences; and deleting a licence cascades down to delating all its transactions. Therefore, all we need to do in the code is delete the invoice and the db will take care of deleting the rest.