Add initial stand-in implementation
We still have a test failing. The reason is the bill run plugin. It is seeing a `PATCH` request so asking the `bill run` if it is editable. We have not included the status of `approved` in its `$editable()` function which is why we're getting an error rather than the 204 we expected.