Add stand-in send bill run endpoint (#245)
https://trello.com/c/kU8RWnKq

This is the first step in adding a `PATCH /v2/{regime}/bill-run/{id}/send` endpoint to the API.

The `/send` endpoint is the means by which a client system confirms a bill run can be sent to SOP for billing. In [Version 1](https://github.com/defra/charging-module-api) it would change the status of the bill run to pending and then a 'behind the scenes' job would generate a transaction file for it, and upload it to AWS S3.

It is also responsible for generating the transaction reference, a key thing for both the client system and SOP.

We're initially just focusing on the reference generation and updating the bill run status and this change covers just creating an initial endpoint. Creating the transaction file will come later.

* Notes

During implementation, we had a test failing. The reason was the bill run plugin. It was seeing a `PATCH` request so asking the `bill run` if it was editable. We had not included the status of `approved` in its `$editable()` function which is why we were getting an error rather than the 204 we expected. Adding 'approved' as an editable state solved the issue.

From the business perspective, it is only when a bill run is 'sent' that it is too late to be edited.