Update BillRunService to expect bill run
With the service no longer fetching the instance of `BillRunModel` we need to pass it in. In normal context this means the controller will pass it into `CreateTransactionService`, which then passes it into `BillRunService`.

Because of how much we lean on `CreateTransactionService` in our tests, we updated the implementation first as it was the easiest way of surfacing which tests needed to be updated.