Update BillRunGenerator in 'test'
We use the bill run generator in our unit tests and expose it via the test endpoint. Behind the scenes, it depends heavily on `CreateTransactionService` which means we needed to update the generator to except a `BillRunModel` instance instead of just the ID.

We also found that it meant we needed to do a fetch in the test controller. The `CreateBillRunService` returns what will be the response to the client system, not an actual bill run instance. We could have amended the service to alter its return value based on a flag, as we've done in `CalculateChargeService`. But this PR has already exploded so we felt it more prudent to take the quick route, and chalk that up to another PR later!