Update CreateTransactionService tests
This also needed to be updated so that we were passing in a `BillRunModel` instance. But this also meant some re-jigging so that we could avoid duplicating the 'create a bill run' call. Finally, we deleted tests that were actually testing the underlying `BillRunService`, especially as it no longer does these tests.