Implement create transaction patch service
The service can take a transaction and generate the appropriate patch for use in a Objection `model.query().patch()` call.

The `_generatePatch()` runs a little long but it is relevatively simple to follow and sticks to 2 conventions

- we update only what needs updating
- we stick to the `raw()` params format, even where we know the value and could just write it into the string

We do have a mix of `raw()` formats elsewhere in the code but the params format (where the value is passed as a param instead of hard coded into the string) is the most flexible. Sticking to one convention should avoid confusion in the future (we'll go back and update the other calls soon!)